Iowa Liquor Sales by Category

Monthly summary of alcohol (class E) sales in Iowa by alcohol category

category_sales

Format

a data frame with 6 variables

year_mon

Month and year of recorded sale information for the respective alcohol category

category_name

Name of category of type of alcohol. This field is used to denote multiple brands that fall into the same category.

state_cost

Cost incurred by State of Iowa to purchase liquor from vendor in US dollars (not normalized)

bottles_sold

Number of bottles sold for the respective year-month

retail_revenue

Sales revenue in US dollars (not normalized)

liters_sold

Volume of liquor sold

Source

State of Iowa Data Portal website